requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
joblib>=1.2.0
pyyaml>=6.0
mlflow>=2.0.0